requirements.txt
python-telegram-bot==12.0.0b1              
pyeasyga==0.3.1
pandas==1.0.3
pymongo
dnspython
